Add PackageSqlite3c sbt task
It is useful to have a jar that just contains the Sqlite3C class because
that is the only class that depends on the native libscalaqlite. This
matters if you are using scalaqlite in a webapp that has hot reloads.
Any class that loads a native library can only be loaded once per jvm.
To handle this some webapps, such as jetty, allow the user to set some
classes to be loaded with a specific classloader. Generally this
requires you to check in a jar in the webapp's lib directory.
Previously, because scalaqlite requires the scala standard library, the
user would have also had to check in the scala standard library into the
lib directory. With this change, they'll only have to check in the tiny
jar generated by this task.